change in error handling and user

+++ JTrackREST/src/main/java/com/jtrack/service/UserService.java
@@ -8,10 +8,14 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jtrack.dao.UserDao;
+import com.jtrack.exception.InvalidDataException;
 import com.jtrack.model.User;
 
 @Service
@@ -38,31 +42,91 @@ public class UserService {
 		return null;
 	}
 	
-	public boolean userExists(String userId) {
-		User userExisting = getUser(userId);
-		return (userExisting != null);
-	}
-	
-	public boolean userValid(String userId) {
-		return (userId != null && !userId.trim().isEmpty());
-	}
-	
-	public User addUser(User user) {
+	public User addUser(User user) throws InvalidDataException {
 		logger.info("addUser({})", user);
+		
+		if(!isCurrentUserAdmin()) {
+			throw new InvalidDataException("No acess to Create User");
+    	}
+		
+		if(userInvalid(user.getUserId())) {
+			throw new InvalidDataException("User is required");
+		}
+		
+		if(pwordInvalid(user.getPword())) {
+			throw new InvalidDataException("Password is required");
+		}
+		
+		if(userExists(user.getUserId())) {
+			throw new InvalidDataException("User already exists");
+		}
+		
+		user.setUserCrt(getCurrentUserId());
 		user.setDateCrt(new Date());
 		 
 	    return userDao.save(user);
 	}
 	
-	public void deleteUser(String userId) {
+	public void deleteUser(String userId) throws InvalidDataException {
 		logger.info("deleteUser({})", userId);
+		
+		if(!isCurrentUserAdmin()) {
+			throw new InvalidDataException("No acess to Delete User");
+    	}
+		
+		if(!userExists(userId)) {
+			throw new InvalidDataException("User does not exist");
+		}
+		
 		userDao.deleteById(userId);
 	}
 	
-	public User updateUser(User user) {
+	public User updateUser(User user) throws InvalidDataException {
 		logger.info("updateUser({})", user);
+		
+		if(!isCurrentUserAdmin()) {
+			throw new InvalidDataException("No acess to Update User");
+    	}
+		
+		if(pwordInvalid(user.getPword())) {
+			throw new InvalidDataException("Password is required");
+		}
+		
+		if(!userExists(user.getUserId())) {
+			throw new InvalidDataException("User does not exist");
+		}
+		
+		user.setUserMod(getCurrentUserId());
 		user.setDateMod(new Date());
 		
 		return userDao.save(user);
+	}
+	
+	public String getCurrentUserId() {
+		String user = "";
+		
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		if (authentication != null && !(authentication instanceof AnonymousAuthenticationToken)) {
+			user = authentication.getName();
+		}
+		
+		return user.toUpperCase();
+	}	
+	
+	private boolean isCurrentUserAdmin() {
+		return "ADMIN".equals(getCurrentUserId());
+	}
+	
+	private boolean userExists(String userId) {
+		User userExisting = getUser(userId);
+		return (userExisting != null);
+	}
+	
+	private boolean userInvalid(String userId) {
+		return (userId == null || userId.trim().isEmpty());
+	}
+	
+	private boolean pwordInvalid(String pword) {
+		return (pword == null || pword.trim().isEmpty());
 	}
 }
